회전 동기화 패킷명 수정

+++ Source/TestProject/Network/Core/TPClient.h
@@ -38,7 +38,7 @@ private:
 	bool isConnected;
 
 	char* serverIp;
-	uint8_t serverPort;
+	uint16_t serverPort;
 	
 	SOCKET hSocket;
 	SOCKADDR_IN recvAddr;
